requirements.txt
Django==2.1.4
django-filter==2.0.0
django-admin-rangefilter==0.3.12
django-cors-headers==2.1.0
djangorestframework==3.9.0
django-storages==1.7.1
Markdown==3.0.1
mysqlclient==1.3.14
pytz==2018.7
Pillow==5.4.1

# Django admin theme
django-jet==1.0.8

# S3 file storage
boto3==1.9.103
botocore==1.12.103
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
s3transfer==0.2.0

# Configuration
django-environ==0.4.4

# Excel uploads
openpyxl==2.5.12
django-excel==0.0.10
xlrd==1.2.0
django-import-export==1.1.0
tablib>=0.12

# Excel, CSV, XLSX, XLS Extract
drf-renderer-xlsx==0.3.2

#  PDF Generation
#  django-easy-pdf==0.2.0
WeasyPrint>=0.34
xhtml2pdf==0.2.3